enabled back bf16, fix layernorm bug, keep fp16 ds config for future

+++ open_clip_mrl/src/open_clip/transformer.py
@@ -15,7 +15,7 @@ class LayerNormFp32(nn.LayerNorm):
 
     def forward(self, x: torch.Tensor):
         orig_type = x.dtype
-        x = F.layer_norm(x.to(torch.float32), self.normalized_shape, self.weight, self.bias, self.eps)
+        x = F.layer_norm(x, self.normalized_shape, self.weight, self.bias, self.eps)
         return x.to(orig_type)
